requirements.txt
# Video Dubbing Tool (Streamlit GUI)
torch>=1.13.0
torchaudio>=0.13.0
numpy==1.26.0
openai-whisper>=20240930
TTS>=0.22.0
moviepy==1.0.3
ffmpeg-python>=0.2.0
soundfile>=0.12.1
librosa>=0.10.0
noisereduce>=2.0.1
scipy>=1.10.0
demucs>=4.0.0
streamlit>=1.20.0
